% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap.synergy_result}
\title{Plot a CI Heatmap}
\usage{
plot_heatmap(x, ...)

\method{plot_heatmap}{synergy_result}(x, ...)
}
\arguments{
\item{x}{A \code{synergy_result} object from \code{analyse_synergy()}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot a CI Heatmap
}
