% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surface.R
\name{plot_surface}
\alias{plot_surface}
\alias{plot_surface.synergy_result}
\title{Plot a 3D Response Surface}
\usage{
plot_surface(x, ...)

\method{plot_surface}{synergy_result}(x, ...)
}
\arguments{
\item{x}{A \code{synergy_result} object from \code{analyse_synergy()}.}

\item{...}{Additional arguments (unused).}
}
\value{
A \code{plotly} object.
}
\description{
Plot a 3D Response Surface
}
