% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.synergy_result}
\alias{print.synergy_result}
\title{Print a Synergy Result}
\usage{
\method{print}{synergy_result}(x, ...)
}
\arguments{
\item{x}{A \code{synergy_result} object from \code{analyse_synergy()}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
\code{x} invisibly.
}
\description{
Print a Synergy Result
}
